+++ src/charmy/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'charmy'
-__version__ = '0.1.4'
-__build__ = 0x000005
+__version__ = '0.1.7'
+__build__ = 0x000008
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2015 Artur Barseghyan'
+__copyright__ = '2015-2016 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'