+++ src/charmy/base.py
@@ -1,9 +1,3 @@
-__title__ = 'charmy.base'
-__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2015 Artur Barseghyan'
-__license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('BaseInstaller',)
-
 import os
 
 from six.moves import configparser
@@ -19,9 +13,15 @@ from .constants import (
 )
 from .helpers import download_file
 
+__title__ = 'charmy.base'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2015-2016 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('BaseInstaller',)
+
+
 class BaseInstaller(object):
-    """
-    Base installer.
+    """Base installer.
 
         - uid (str): Unique identifier of the installer.
         - platform (str): Platform.
@@ -41,8 +41,8 @@ class BaseInstaller(object):
                                          which would sym-link to latest
                                          PyCharm executable.
         - db_ready (bool): Indicates whether database is ready.
-
     """
+
     os = None
     platform = None
     installation_dir = None
@@ -54,7 +54,7 @@ class BaseInstaller(object):
     db_ready = False # Indicates whether database is ready
 
     def __init__(self, *args, **kwargs):
-        """
+        """Constructor.
 
         :param args:
         :param kwargs:
@@ -78,6 +78,7 @@ class BaseInstaller(object):
         self._sync_db()
 
     def _create_dirs(self):
+        """Create dirs (internal method)."""
         dirs = (
             self.installation_dir,
             self.config_dir,
@@ -90,13 +91,15 @@ class BaseInstaller(object):
         return True
 
     def _create_files(self):
+        """Create files (internal method)."""
         if not os.path.isfile(self.config_ini_filename):
             config_ini_file = open(self.config_ini_filename, 'w')
             config_ini_file.close()
 
     def _prepare_db(self):
-        """
-        Prepare the database (create if not exist). Prepare the models.
+        """Prepare the database (create if not exist).
+
+        Also, prepare the models.
 
         :return:
         """
@@ -109,8 +112,7 @@ class BaseInstaller(object):
         self.db_ready = True
 
     def drop_db(self):
-        """
-        Drops the database.
+        """Drop the database.
 
         :return:
         """
@@ -118,17 +120,15 @@ class BaseInstaller(object):
             os.remove(self.db_name)
 
     def _get_distribution_class(self):
-        """
-        Gets the `Distribution` model for saving installed distributions.
+        """Get the `Distribution` model for saving installed distributions.
 
         :return:
         """
         db = self.db
 
         class Distribution(db.Model):
-            """
-            Distribution model for saving installed distributions.
-            """
+            """Distribution model for saving installed distributions."""
+
             id = db.Column(db.Integer, primary_key=True)
             version = db.Column(db.String(80))
             edition = db.Column(db.String(120))
@@ -145,12 +145,14 @@ class BaseInstaller(object):
             )
 
             def __init__(self, version, edition, path, active):
+                """Constructor."""
                 self.version = version
                 self.edition = edition
                 self.path = path
                 self.active = active
 
             def __repr__(self):
+                """Representation."""
                 return '<Distribution {0} {1}>'.format(self.version,
                                                        self.edition,
                                                        self.path,
@@ -159,8 +161,8 @@ class BaseInstaller(object):
         return Distribution
 
     def _sync_db(self):
-        """
-        Creates database if not exist.
+        """Create database if not exist.
+
         :return:
         """
         if not self.db_ready:
@@ -168,10 +170,11 @@ class BaseInstaller(object):
         self.db.create_all()
 
     def download(self, version, edition, use_cache=True):
-        """
-        Downloads the PyCharm
+        """Download the PyCharm.
+
         :param version:
         :param edition:
+        :param bool use_cache:
         :return:
         """
         assert edition in EDITIONS
@@ -181,6 +184,7 @@ class BaseInstaller(object):
         return downloaded_file
 
     def _build_download_link(self, version, edition):
+        """Build download link."""
         return DOWNLOAD_LINK_PATTERN.format(
             **{
                 'edition': edition,
@@ -190,8 +194,7 @@ class BaseInstaller(object):
         )
 
     def _install(self, version, edition, use_cache=True, destination=None):
-        """
-        Installs the downloaded PyCharm file.
+        """Install the downloaded PyCharm file (internal method).
 
         :param str version:
         :param str edition:
@@ -218,7 +221,8 @@ class BaseInstaller(object):
             return (False, installation_dir, '')
 
     def install(self, file, destination=None):
-        """
+        """Install the downloaded PyCharm.
+
         This method should return True on success. Otherwise considered to
         be failed.
 
@@ -228,7 +232,7 @@ class BaseInstaller(object):
         raise NotImplemented
 
     def _write_destination_info_to_config_ini(self, destination):
-        """
+        """Write destination info to the config.ini.
 
         :param str destination:
         :return:
@@ -241,8 +245,7 @@ class BaseInstaller(object):
         config_ini_file.close()
 
     def _read_destination_from_config_ini(self):
-        """
-        Reads destination from config ini.
+        """Read destination from config ini.
 
         :return str:
         """
@@ -262,9 +265,9 @@ class BaseInstaller(object):
 
     def _write_distribution_info_to_db(self, version, edition,
                                        installation_dir):
-        """
-        Writes distribution data into the database. Returns True on success
-        and False on duplicates.
+        """Write distribution data into the database.
+
+        Return True on success and False on duplicates.
 
         :param str version:
         :param str edition:
@@ -277,7 +280,6 @@ class BaseInstaller(object):
             .filter_by(path=installation_dir, active=True) \
             .update({"active": False})
         self.db.session.commit()
-
 
         # Now save the distribution.
         distribution = self.distribution_cls(version,
@@ -292,15 +294,16 @@ class BaseInstaller(object):
             self.db.session.rollback()
             self.distribution_cls \
                 .query \
-                .filter_by(version=version, edition=edition, path=installation_dir) \
+                .filter_by(version=version,
+                           edition=edition,
+                           path=installation_dir) \
                 .update({"active": True})
             self.db.session.commit()
 
             return False
 
     def _versions(self):
-        """
-        Gets installed versions.
+        """List installed versions (internal method).
 
         :return:
         """
@@ -311,10 +314,11 @@ class BaseInstaller(object):
                 for dist in self.distribution_cls.query.all()]
 
     def versions(self):
+        """List installed versions."""
         return self._versions()
 
     def _uninstall(self, version, edition):
-        """
+        """Uninstall given version/edition (internal method).
 
         :param version:
         :param edition:
@@ -323,9 +327,11 @@ class BaseInstaller(object):
         return self.uninstall(version, edition)
 
     def uninstall(self, version, edition):
+        """Uninstall given version/edition."""
         raise NotImplemented
 
     def _activate(self, version, edition):
+        """Activate given version/edition (internal method)."""
         activated, installation_dir, message = self.activate(version, edition)
         if activated:
             self._write_distribution_info_to_db(version,
@@ -334,7 +340,8 @@ class BaseInstaller(object):
         return (activated, installation_dir, message)
 
     def activate(self, version, edition):
-        """
+        """Activate given version/edition.
+
         This method should return a tuple (True, destination, message) on
         success. Otherwise considered to be failed.