+++ src/charmy/utils.py
@@ -1,9 +1,3 @@
-__title__ = 'charmy.utils'
-__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2015 Artur Barseghyan'
-__license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('get_installer',)
-
 import platform
 
 from .constants import PLATFORM_LINUX
@@ -11,8 +5,14 @@ from .installers.linux.ubuntu import UID as LINUX_UBUNTU_UID
 from .installers.linux.ubuntu.installer import UbuntuInstaller
 from .installers.linux.other.installer import OtherLinuxInstaller
 
+__title__ = 'charmy.utils'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2015-2016 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('get_installer',)
+
 def get_installer():
-    """
+    """Get installer.
 
     :return:
     """