+++ CHANGELOG.rst
@@ -1,5 +1,26 @@
 Changelog
 =========
+0.1.7
+-----
+2016-09-09
+
+- Minor fixes.
+
+0.1.6
+-----
+2016-09-09
+
+- Add `discover-installed-versions` command.
+- Add `download` command.
+- Fixed latest version detection (since it stopped working recently due to
+  changes on PyCharm website).
+
+0.1.5
+-----
+2015-10-21
+
+- Fixes in installer regarding installation of professional version.
+
 0.1.4
 -----
 2015-10-10